in process

+++ main/src/main/java/ru/practicum/service/CategoryService.java
@@ -4,10 +4,12 @@ import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import ru.practicum.dto.NewCategoryDto;
+import ru.practicum.exception.ConflictExeption;
 import ru.practicum.exception.NotFoundException;
 import ru.practicum.mappers.CategoryMapper;
 import ru.practicum.model.Category;
 import ru.practicum.repository.CategoryRepository;
+import ru.practicum.repository.EventRepository;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -17,6 +19,7 @@ import java.util.stream.Collectors;
 public class CategoryService {
 
     private final CategoryRepository categoryRepository;
+    private final EventRepository eventRepository;
 
     public List<Category> getCategories(Integer from, Integer size) {
         Integer resultFrom = from.equals(0) ? 1 : from - 1;
@@ -35,6 +38,10 @@ public class CategoryService {
         if (!categoryRepository.existsById(id)) {
             throw new NotFoundException("Category with id=" + id + " was not found");
         }
+        if (eventRepository.countEventsWithCategory(id) > 0) {
+            throw new ConflictExeption("The category is not empty");
+        }
+
         categoryRepository.deleteById(id);
     }
